--- pkg/apis/internal.kusion.io/group.go
@@ -1,3 +1,0 @@
-package internalkusionio
-
-const Group = "internal.kusion.io"
