+++ pkg/backend/storages/completion.go
@@ -3,7 +3,7 @@ package storages
 import (
 	"os"
 
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/util/kfile"
 )
