+++ pkg/backend/storages/oss.go
@@ -3,7 +3,7 @@ package storages
 import (
 	"github.com/aliyun/aliyun-oss-go-sdk/oss"
 
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/engine/state"
 	statestorages "kusionstack.io/kusion/pkg/engine/state/storages"
 	"kusionstack.io/kusion/pkg/workspace"
