+++ pkg/backend/storages/validation.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 )
 
 var (
