+++ pkg/cmd/apply/apply_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	apiv1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
-	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	v1 "kusionstack.io/kusion/pkg/apis/status/v1"
 	"kusionstack.io/kusion/pkg/backend/storages"
 	"kusionstack.io/kusion/pkg/cmd/generate"
@@ -55,7 +54,7 @@ var (
 )
 
 func NewApplyOptions() *ApplyOptions {
-	storageBackend := storages.NewLocalStorage(&internalv1.BackendLocalConfig{
+	storageBackend := storages.NewLocalStorage(&apiv1.BackendLocalConfig{
 		Path: filepath.Join("", "state.yaml"),
 	})
 	return &ApplyOptions{
