+++ pkg/cmd/preview/preview_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	apiv1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
-	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	v1 "kusionstack.io/kusion/pkg/apis/status/v1"
 	"kusionstack.io/kusion/pkg/backend/storages"
 	"kusionstack.io/kusion/pkg/cmd/generate"
@@ -59,7 +58,7 @@ var (
 )
 
 func NewPreviewOptions() *PreviewOptions {
-	storageBackend := storages.NewLocalStorage(&internalv1.BackendLocalConfig{
+	storageBackend := storages.NewLocalStorage(&apiv1.BackendLocalConfig{
 		Path: filepath.Join("", "state.yaml"),
 	})
 	return &PreviewOptions{
