+++ pkg/config/util.go
@@ -1,7 +1,7 @@
 package config
 
 import (
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 )
 
 // GetConfig returns the structured config stored in the config file. The validation of the config is not checked.
