+++ pkg/domain/entity/backend.go
@@ -3,7 +3,7 @@ package entity
 import (
 	"time"
 
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/domain/constant"
 )
