+++ pkg/domain/request/backend_request.go
@@ -1,6 +1,6 @@
 package request
 
-import v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+import v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 
 // CreateBackendRequest represents the create request structure for
 // backend.
