+++ pkg/engine/api/builders/appconfig_builder.go
@@ -21,13 +21,12 @@ import (
 	"kcl-lang.io/kpm/pkg/api"
 
 	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
-	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/modules"
 	"kusionstack.io/kusion/pkg/modules/generators"
 )
 
 type AppsConfigBuilder struct {
-	Apps      map[string]internalv1.AppConfiguration
+	Apps      map[string]v1.AppConfiguration
 	Workspace *v1.Workspace
 }
 
@@ -37,7 +36,7 @@ func (acg *AppsConfigBuilder) Build(kclPackage *api.KclPackage, project *v1.Proj
 	}
 
 	var gfs []modules.NewGeneratorFunc
-	err := modules.ForeachOrdered(acg.Apps, func(appName string, app internalv1.AppConfiguration) error {
+	err := modules.ForeachOrdered(acg.Apps, func(appName string, app v1.AppConfiguration) error {
 		if kclPackage == nil {
 			return fmt.Errorf("kcl package is nil when generating app configuration for %s", appName)
 		}
