+++ pkg/infra/persistence/backend_model.go
@@ -1,7 +1,7 @@
 package persistence
 
 import (
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/domain/entity"
 
 	"gorm.io/gorm"
