+++ pkg/infra/persistence/backend_test.go
@@ -7,7 +7,8 @@ import (
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/stretchr/testify/require"
 	"gorm.io/gorm"
-	v1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
+
+	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/domain/entity"
 )
