+++ pkg/modules/generators/app_configurations_generator.go
@@ -13,7 +13,6 @@ import (
 	"kcl-lang.io/kpm/pkg/package"
 
 	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
-	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/log"
 	"kusionstack.io/kusion/pkg/modules"
 	"kusionstack.io/kusion/pkg/modules/generators/workload"
@@ -25,7 +24,7 @@ type appConfigurationGenerator struct {
 	project      string
 	stack        string
 	appName      string
-	app          *internalv1.AppConfiguration
+	app          *v1.AppConfiguration
 	ws           *v1.Workspace
 	dependencies *pkg.Dependencies
 }
@@ -38,7 +37,7 @@ func NewAppConfigurationGenerator(
 	project string,
 	stack string,
 	appName string,
-	app *internalv1.AppConfiguration,
+	app *v1.AppConfiguration,
 	ws *v1.Workspace,
 	dependencies *pkg.Dependencies,
 ) (modules.Generator, error) {
@@ -80,7 +79,7 @@ func NewAppConfigurationGeneratorFunc(
 	project string,
 	stack string,
 	appName string,
-	app *internalv1.AppConfiguration,
+	app *v1.AppConfiguration,
 	ws *v1.Workspace,
 	kpmDependencies *pkg.Dependencies,
 ) modules.NewGeneratorFunc {
@@ -150,7 +149,7 @@ func (g *appConfigurationGenerator) Generate(spec *v1.Spec) error {
 	return nil
 }
 
-func PatchWorkload(workload *v1.Resource, patcher *internalv1.Patcher) error {
+func PatchWorkload(workload *v1.Resource, patcher *v1.Patcher) error {
 	if patcher == nil {
 		return nil
 	}
@@ -251,12 +250,12 @@ func PatchWorkload(workload *v1.Resource, patcher *internalv1.Patcher) error {
 
 // moduleConfig represents the configuration of a module, either devConfig or platformConfig can be nil
 type moduleConfig struct {
-	devConfig      internalv1.Accessory
+	devConfig      v1.Accessory
 	platformConfig v1.GenericConfig
 	ctx            v1.GenericConfig
 }
 
-func (g *appConfigurationGenerator) callModules(projectModuleConfigs map[string]v1.GenericConfig) (resources []v1.Resource, patchers []internalv1.Patcher, err error) {
+func (g *appConfigurationGenerator) callModules(projectModuleConfigs map[string]v1.GenericConfig) (resources []v1.Resource, patchers []v1.Patcher, err error) {
 	pluginMap := make(map[string]*modules.Plugin)
 	defer func() {
 		for _, plugin := range pluginMap {
@@ -324,7 +323,7 @@ func (g *appConfigurationGenerator) callModules(projectModuleConfigs map[string]
 
 		// parse patcher
 		for _, patcher := range response.Patchers {
-			temp := &internalv1.Patcher{}
+			temp := &v1.Patcher{}
 			err = yaml.Unmarshal(patcher, temp)
 			if err != nil {
 				return nil, nil, err
@@ -364,7 +363,7 @@ func (g *appConfigurationGenerator) buildModuleConfigIndex(platformModuleConfigs
 	return indexModuleConfig, nil
 }
 
-func parseModuleKey(accessory internalv1.Accessory, dependencies *pkg.Dependencies) (string, error) {
+func parseModuleKey(accessory v1.Accessory, dependencies *pkg.Dependencies) (string, error) {
 	split := strings.Split(accessory["_type"].(string), ".")
 	moduleName := split[0]
 	// find module namespace and version
