+++ pkg/modules/generators/app_configurations_generator_test.go
@@ -16,7 +16,6 @@ import (
 	pkg "kcl-lang.io/kpm/pkg/package"
 
 	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
-	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/modules"
 	"kusionstack.io/kusion/pkg/modules/proto"
 	jsonutil "kusionstack.io/kusion/pkg/util/json"
@@ -160,16 +159,16 @@ func TestNewAppConfigurationGeneratorFunc(t *testing.T) {
 	})
 }
 
-func buildMockApp() (string, *internalv1.AppConfiguration) {
-	return "app1", &internalv1.AppConfiguration{
-		Workload: &internalv1.Workload{
-			Header: internalv1.Header{
-				Type: internalv1.TypeService,
+func buildMockApp() (string, *v1.AppConfiguration) {
+	return "app1", &v1.AppConfiguration{
+		Workload: &v1.Workload{
+			Header: v1.Header{
+				Type: v1.TypeService,
 			},
-			Service: &internalv1.Service{
-				Base: internalv1.Base{},
+			Service: &v1.Service{
+				Base: v1.Base{},
 				Type: "Deployment",
-				Ports: []internalv1.Port{
+				Ports: []v1.Port{
 					{
 						Port:     80,
 						Protocol: "TCP",
@@ -291,7 +290,7 @@ func Test_patchWorkload(t *testing.T) {
 	}
 
 	t.Run("Patch labels and annotations", func(t *testing.T) {
-		patcher := &internalv1.Patcher{
+		patcher := &v1.Patcher{
 			Labels:      map[string]string{"newLabel": "newValue"},
 			Annotations: map[string]string{"newAnnotation": "newValue"},
 		}
@@ -320,7 +319,7 @@ func Test_patchWorkload(t *testing.T) {
 	})
 
 	t.Run("Patch environment variables", func(t *testing.T) {
-		patcher := &internalv1.Patcher{
+		patcher := &v1.Patcher{
 			Environments: []corev1.EnvVar{
 				{
 					Name:  "NEW_ENV",
