+++ pkg/modules/generators/workload/secret/secret_generator.go
@@ -13,7 +13,6 @@ import (
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 
 	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
-	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/modules"
 	"kusionstack.io/kusion/pkg/secrets"
 )
@@ -21,7 +20,7 @@ import (
 type secretGenerator struct {
 	project         string
 	namespace       string
-	secrets         map[string]internalv1.Secret
+	secrets         map[string]v1.Secret
 	secretStoreSpec *v1.SecretStoreSpec
 }
 
@@ -31,7 +30,7 @@ type GeneratorRequest struct {
 	// Namespace represents the K8s Namespace
 	Namespace string
 	// Workload represents the Workload configuration
-	Workload *internalv1.Workload
+	Workload *v1.Workload
 	// SecretStoreSpec contains configuration to describe target secret store.
 	SecretStoreSpec *v1.SecretStoreSpec
 }
@@ -41,7 +40,7 @@ func NewSecretGenerator(request *GeneratorRequest) (modules.Generator, error) {
 		return nil, fmt.Errorf("project name must not be empty")
 	}
 
-	var secretMap map[string]internalv1.Secret
+	var secretMap map[string]v1.Secret
 	if request.Workload.Service != nil {
 		secretMap = request.Workload.Service.Secrets
 	} else {
@@ -91,7 +90,7 @@ func (g *secretGenerator) Generate(spec *v1.Spec) error {
 // generateSecret generates target secret based on secret type. Most of these secret types are just semantic wrapper
 // of native Kubernetes secret types:https://kubernetes.io/docs/concepts/configuration/secret/#secret-types, and more
 // detailed usage info can be found in public documentation.
-func (g *secretGenerator) generateSecret(secretName string, secretRef internalv1.Secret) (*corev1.Secret, error) {
+func (g *secretGenerator) generateSecret(secretName string, secretRef v1.Secret) (*corev1.Secret, error) {
 	switch secretRef.Type {
 	case "basic":
 		return g.generateBasic(secretName, secretRef)
@@ -110,7 +109,7 @@ func (g *secretGenerator) generateSecret(secretName string, secretRef internalv1
 
 // generateBasic generates secret used for basic authentication. The basic secret type
 // is used for username / password pairs.
-func (g *secretGenerator) generateBasic(secretName string, secretRef internalv1.Secret) (*corev1.Secret, error) {
+func (g *secretGenerator) generateBasic(secretName string, secretRef v1.Secret) (*corev1.Secret, error) {
 	secret := initBasicSecret(g.namespace, secretName, corev1.SecretTypeBasicAuth, secretRef.Immutable)
 	secret.Data = grabData(secretRef.Data, corev1.BasicAuthUsernameKey, corev1.BasicAuthPasswordKey)
 
@@ -126,7 +125,7 @@ func (g *secretGenerator) generateBasic(secretName string, secretRef internalv1.
 
 // generateToken generates secret used for password. Token secrets are useful for generating
 // a password or secure string used for passwords when the user is already known or not required.
-func (g *secretGenerator) generateToken(secretName string, secretRef internalv1.Secret) (*corev1.Secret, error) {
+func (g *secretGenerator) generateToken(secretName string, secretRef v1.Secret) (*corev1.Secret, error) {
 	secret := initBasicSecret(g.namespace, secretName, corev1.SecretTypeOpaque, secretRef.Immutable)
 	secret.Data = grabData(secretRef.Data, "token")
 
@@ -139,7 +138,7 @@ func (g *secretGenerator) generateToken(secretName string, secretRef internalv1.
 }
 
 // generateOpaque generates secret used for arbitrary user-defined data.
-func (g *secretGenerator) generateOpaque(secretName string, secretRef internalv1.Secret) (*corev1.Secret, error) {
+func (g *secretGenerator) generateOpaque(secretName string, secretRef v1.Secret) (*corev1.Secret, error) {
 	secret := initBasicSecret(g.namespace, secretName, corev1.SecretTypeOpaque, secretRef.Immutable)
 	secret.Data = grabData(secretRef.Data, maps.Keys(secretRef.Data)...)
 	return secret, nil
@@ -147,8 +146,8 @@ func (g *secretGenerator) generateOpaque(secretName string, secretRef internalv1
 
 // generateCertificate generates secret used for storing a certificate and its associated key.
 // One common use for TLS Secrets is to configure encryption in transit for an Ingress, but
-// you can also use it with other resources or directly in your internalv1.
-func (g *secretGenerator) generateCertificate(secretName string, secretRef internalv1.Secret) (*corev1.Secret, error) {
+// you can also use it with other resources or directly in your v1.
+func (g *secretGenerator) generateCertificate(secretName string, secretRef v1.Secret) (*corev1.Secret, error) {
 	secret := initBasicSecret(g.namespace, secretName, corev1.SecretTypeTLS, secretRef.Immutable)
 	secret.Data = grabData(secretRef.Data, corev1.TLSCertKey, corev1.TLSPrivateKeyKey)
 	return secret, nil
@@ -156,7 +155,7 @@ func (g *secretGenerator) generateCertificate(secretName string, secretRef inter
 
 // generateSecretWithExternalProvider retrieves target sensitive information from external secret provider and
 // generates corresponding Kubernetes Secret object.
-func (g *secretGenerator) generateSecretWithExternalProvider(secretName string, secretRef internalv1.Secret) (*corev1.Secret, error) {
+func (g *secretGenerator) generateSecretWithExternalProvider(secretName string, secretRef v1.Secret) (*corev1.Secret, error) {
 	if g.secretStoreSpec == nil {
 		return nil, errors.New("secret store is missing, please add valid secret store spec in workspace")
 	}
